refactor codex output parsers
