+++ src/codex.ts
@@ -4,6 +4,7 @@ import path from "node:path";
 import { extractIssueJournalHandoff } from "./journal";
 import { runCommand } from "./command";
 import { buildCodexConfigOverrideArgs, resolveCodexExecutionPolicy } from "./codex-policy";
+export { extractBlockedReason, extractFailureSignature, extractStateHint } from "./codex-output-parser";
 import { ExternalReviewMissContext, type ExternalReviewMissPattern } from "./external-review-misses";
 import {
   BlockedReason,
@@ -41,72 +42,6 @@ const LIVE_BLOCKER_HANDOFF_SUPPRESSION_STATES = new Set<RunState>([
 ]);
 
 const COMPACT_RESUME_PROMPT_STATES = new Set<RunState>(["planning", "reproducing", "implementing", "stabilizing", "draft_pr"]);
-
-export function extractStateHint(message: string): RunState | null {
-  const match = message.match(/State hint:\s*([a-z_]+)/i);
-  if (!match) {
-    return null;
-  }
-
-  const value = match[1].toLowerCase() as RunState;
-  const supported: RunState[] = [
-    "queued",
-    "planning",
-    "reproducing",
-    "implementing",
-    "local_review_fix",
-    "stabilizing",
-    "draft_pr",
-    "local_review",
-    "pr_open",
-    "repairing_ci",
-    "resolving_conflict",
-    "waiting_ci",
-    "addressing_review",
-    "ready_to_merge",
-    "merging",
-    "done",
-    "blocked",
-    "failed",
-  ];
-
-  return supported.includes(value) ? value : null;
-}
-
-export function extractBlockedReason(message: string): BlockedReason {
-  const match = message.match(/Blocked reason:\s*([a-z_]+)/i);
-  if (!match) {
-    return null;
-  }
-
-  const value = match[1].toLowerCase() as BlockedReason;
-  const supported: BlockedReason[] = [
-    "requirements",
-    "permissions",
-    "secrets",
-    "verification",
-    "manual_review",
-    "manual_pr_closed",
-    "handoff_missing",
-    "unknown",
-    null,
-  ];
-  return supported.includes(value) ? value : null;
-}
-
-export function extractFailureSignature(message: string): string | null {
-  const match = message.match(/Failure signature:\s*(.+)/i);
-  if (!match) {
-    return null;
-  }
-
-  const value = match[1].trim();
-  if (!value || value.toLowerCase() === "none") {
-    return null;
-  }
-
-  return value.slice(0, 500);
-}
 
 export function shouldUseCompactResumePrompt(state: RunState): boolean {
   return COMPACT_RESUME_PROMPT_STATES.has(state);
